UNG-1865 - Moving E2E tests to seperate package
