+++ src/test/scala/com/ubirch/e2e/controllers/SuperAdminControllerSpec.scala
@@ -1,14 +1,17 @@
-package com.ubirch.controllers
+package com.ubirch.e2e.controllers
 
+import com.ubirch.controllers.SuperAdminController
+import com.ubirch.e2e.InjectorHelperImpl
 import com.ubirch.services.jwt.PublicKeyPoolService
-import com.ubirch.{ Awaits, ExecutionContextsTests, FakeTokenCreator, InjectorHelperImpl }
+import com.ubirch.{Awaits, ExecutionContextsTests, FakeTokenCreator}
 import io.prometheus.client.CollectorRegistry
 import org.scalatest.BeforeAndAfterEach
 import org.scalatra.test.scalatest.ScalatraWordSpec
 
 import scala.concurrent.duration.DurationInt
 
-class SuperAdminControllerSpec extends ScalatraWordSpec
+class SuperAdminControllerSpec
+  extends ScalatraWordSpec
   with BeforeAndAfterEach
   with ExecutionContextsTests
   with Awaits {
@@ -19,7 +22,8 @@ class SuperAdminControllerSpec extends ScalatraWordSpec
     "fail when token is not provided" in {
       get("/initialTest") {
         status should equal(401)
-        assert(body == """{"version":"1.0","ok":false,"errorType":"AuthenticationError","errorMessage":"Unauthenticated"}""")
+        assert(
+          body == """{"version":"1.0","ok":false,"errorType":"AuthenticationError","errorMessage":"Unauthenticated"}""")
       }
     }
 
@@ -36,11 +40,11 @@ class SuperAdminControllerSpec extends ScalatraWordSpec
     CollectorRegistry.defaultRegistry.clear()
   }
 
-  protected override def afterAll(): Unit = {
+  override protected def afterAll(): Unit = {
     super.afterAll()
   }
 
-  protected override def beforeAll(): Unit = {
+  override protected def beforeAll(): Unit = {
 
     CollectorRegistry.defaultRegistry.clear()
