+++ src/test/scala/com/ubirch/e2e/keycloak/roles/KeycloakRolesServiceTest.scala
@@ -1,8 +1,8 @@
-package com.ubirch.services.keycloak.roles
-
-import com.ubirch.E2ETestBase
+package com.ubirch.e2e.keycloak.roles
 import com.ubirch.data.KeycloakTestData
+import com.ubirch.e2e.E2ETestBase
 import com.ubirch.models.keycloak.roles.{RoleAlreadyExists, RoleName}
+import com.ubirch.services.keycloak.roles.KeycloakRolesService
 
 import scala.concurrent.duration.DurationInt
