+++ src/test/scala/com/ubirch/e2e/keycloak/users/KeycloakUserServiceTest.scala
@@ -1,9 +1,9 @@
-package com.ubirch.services.keycloak.users
-
-import com.ubirch.E2ETestBase
+package com.ubirch.e2e.keycloak.users
 import com.ubirch.data.KeycloakTestData
+import com.ubirch.e2e.E2ETestBase
 import com.ubirch.models.keycloak.user.UserAlreadyExists
 import com.ubirch.models.user.UserName
+import com.ubirch.services.keycloak.users.KeycloakUserService
 import org.scalactic.StringNormalizations._
 
 import scala.concurrent.duration.DurationInt
